Put both generic widgets in new shared file
